requirements.txt
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.0
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
h11==0.14.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.3.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.12.3
itemadapter==0.9.0
itemloaders==1.3.1
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.24.4
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
parsel==1.9.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20240706
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
Protego==0.3.1
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyOpenSSL==24.2.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pywin32==306
pyzmq==26.1.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.11.2
selenium==4.23.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tldextract==5.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
w3lib==2.2.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zipp==3.20.0
zope.interface==7.0.1
PyMuPDF
